requirements.txt
streamlit
google-generativeai
langchain
langchain-community
pypdf
numpy